loyaltyDiscount + tests fe
